frontend/requirements.txt
fastapi
uvicorn
mangum
python-multipart
sqlalchemy
asyncpg
greenlet
pydantic
pydantic-settings
google-generativeai
python-dotenv
pgvector
# Multi-format processing
openpyxl
pandas
pyarrow
